+++ src/Main.java
@@ -4,7 +4,7 @@ public class Main{
 	public static void main(String[] args) {
 		//Dados do Usuario
 		String chave = "ZZZ";
-		Tabela indice = new Tabela("C:\\Users\\User\\Downloads\\words.txt");
+		IndiceHash indice = new IndiceHash("C:\\Users\\User\\Downloads\\words.txt");
 		//Especificar caminho do arquivo na máquina
 		try {
 			System.out.println(indice.chaveBusca(chave));
